Makefile: Explicitly override Git variables for old bundled Git
For the old bundled Git version we're not overriding the variables
`GIT_VERSION`, `GIT_PATCHES` et al, but instead we were simply reusing
those defined at the global scope. But given that the default version of
our Git distribution is about to change this doesn't work anymore.

Explicitly override the version of bundled Git to prepare for this
change.