Instrument verification worker for observability
This commit adds metrics on the verification worker so the process
becomes more observable. The metrics track the number of dequeued
jobs per storage and then number of completed jobs per storage with
their results. The number of stale leases releases is also tracked.
This gives insight into how the verification work is progressing.