command: Adding ReleaseCgrouResources() function
In a future commit, we will add functionality to the cgroups to be able
to have per-git command cgroups. These will be handled in memory, which
means when a command finishes it will need to tell the cgroups manager
it has finished.

Add a function that takes a function as a parameter. This allows for
maximum flexibility in how to release cgroup resources its using.