replicator: Ignore maintenance-style replication jobs
Starting with Gitaly v14.10, we use a best-effort strategy to route
maintenance-style RPCs. With this strategy, we always route such RPC
calls to all healthy nodes which have the given target repository,
regardless of whether they're currently consistent or not. Under this
new schema we also don't create any replication jobs anymore: we deem
maintenance to not be crucial enough to warrant all the infrastructure
here, and instead we can just wait for the next maintenance call to
optimize the repository.

Even though we don't create any replication jobs anymore, we still
handle them right now. And this is a strict requirement for now for
zero-downtime upgrades: a concurrently running Gitaly node may still be
generating maintenance-style replication events, and consequentially we
must still know how to handle them or otherwise we'll start generating
errors.

Despite that, we can still remove all the logic supporting these events
and just make them a no-op. Like this, we make sure to drain the queue
while keeping intact the new best-effort semantics. And furthermore, it
allows us to get rid of a significant chunk of code.

Do so and remove the infrastructure to handle maintenance-style
replication jobs. The no-op can then be removed in the next release
together with a SQL migration which drains all remaining events.