Makefile: Drop bundled Git v2.33.1.gl3
We have finished the migration to bundled Git v2.35.1.gl1 in v14.10. Due
to concerns with zero-downtime upgrades we couldn't yet remove the old
version though. But now that we have waited for a release we can finally
remove the old version.

Remove the infrastructure to build and install bundled Git v2.33.1.gl3.

Changelog: removed