rspec: Convert test repo helper to use newer bundled Git version
The test repository helper in our rspecs is used to create temporary
repositories for testing purposes. The Git version it uses for this
purpose depends on a set of environment variables, but in case it runs
with bundled Git it needs to some more bootstrapping of an execution
environment in order to be able to properly clone the repository.

The bundled Git version is currently hard-coded to `gitaly-git`, which
we're about to phase out in favor of `gitaly-git-v2.35.1.gl1`. With the
removal of the old version the helper is thus about to break.

Fix this by switching over to the newer bundled Git version. While we
could do fancy things like injecting the version or auto-detecting it,
it doesn't really feel worth it over simply hard-coding it. We're
ultimately going to retire the sidecar over the next few releases, so we
shouldn't spend more time than necessary on maintaining it.