git: Open-code execution of git-version(1)
Due to a new incompatibility in Git with regards to the newly introduced
`core.fsync` option we'll have to inject different global options based
on the Git version. Consequentially, we are forced to detect the Git
version when spawning any Git commands. This creates a chicken-and-egg
problem though: we need to spawn a command via `newCommand()`, which in
turn tries to detect the Git version by running `newCommand()` to run
git-version(1).

Prepare for this upcoming version check by open-coding `GitVersion()` so
that it doesn't use `newCommand()` anymore. Instead, we simply assemble
the arguments ourselves and then use `command.New()`. This gets rid of
the cyclic dependency and paves the path for the new version check.