git: Fix fsyncing config incompatibility with Git v2.36.0
Git v2.36.0 has introduced new infrastructure to allow administrators
more fine-grained control over what files Git will synchronize to disks.
This new infrastructure replaces the old `core.fsyncObjectFiles` config,
which we are currently injecting unconditionally. Unfortunately, the new
Git version has started to emit a warning in that case, which would both
be presented to the user and furthermore it would break a bunch of our
tests.

We're thus forced to conditionally inject either the old or the new
config depending on which Git version we're running. Let's do so by
injecting `core.fsync=objects,derived-metadata` instead when running Git
v2.36.0 or newer. This has the same semantics as what we had previously
with `core.fsyncObjectFiles=true`:

    - `default` is what Git did by default previously, and expands to
      `objects,derived-metadata,-loose-object`. Note that the current
      documentation in gitconfig(1) is wrong as it pretends that it
      instead expands to `committed,-loose-object`.

    - Because we want to also synchronize loose objects though we can
      drop the `-loose-object` part.

This fixes compatibility witih v2.36.0 while continuing to do the same
as we did before.

Changelog: fixed