git: Fix corruption of refs on hard-resets
In the past, we had several incidents where repositories ended up with
corrupted loose references immediately after a server was experiencing a
hard shut-down. The root cause of this is that Git historically didn't
flush the contents of loose references to disk before renaming them into
place. This usage is explicitly documented to be invalid by various
filesystems in the Linux kernel. Most importantly, documentation of the
ext4 filesystem says that:

    Many broken applications don't use fsync() when replacing existing
    files via patterns such as fd = open("foo.new")/write(fd,..)/close(fd)/
    rename("foo.new", "foo"), or worse yet, fd = open("foo",
    O_TRUNC)/write(fd,..)/close(fd).

With Git v2.36.0 new infrastructure was introduced that provides finer
grained control of exactly what data should be flushed to disk via the
new `core.fsync` option. As part of that effort we have also upstreamed
patches which give us the ability to also flush references to disk
before they're renamed into place.

So let's finally fix this source of corruption on newer Git versions by
enabling fsync(3P) for references.

Changelog: fixed