gittest: Refactor intercepting command factory options to be extensible
It's currently only possible to pass options intended for the real Git
command factory when constructing an intercepting factory. We're about
to add another option though that changes behaviour of the intercepting
factory itself.

Introduce a new `InterceptingCommandFactoryOption` type to make this
more flexible and allow for the new usecase.