catfile: Deduplicate logic to tear down processes
Both the `CatfileReader` and `CatfileInfoReader` have the same logic to
tear down the process when the context gets cancelled. This logic can
easily be unified into the cache: it already got a Goroutine that waits
for the context to get cancelled to decrement the process counter, so it
is a perfect fit to also host the other logic.

Refactor the code to do this. While this reduces complexity, it also
cuts the number of Goroutines in half.