commit: Avoid reallocating result array when sending object existence
The sender we're using for chunked responses in `CheckObjectsExists()`
is reallocating the result array every time we have sent out a message.
This is needlessly wasteful though: we can just truncate the array and
thus avoid a memory allocation. It's likely not going to matter much in
this context, but the end result is easier to read, too.