Update process and readme by removing module version
The PROCESS.md contains a section on upgrading the module during a major
version upgrade. This is no longer needed as we're changing the package
name to not include the version number. Remove this section.

The README.md also contains references to the major version number in
the package name. Replace the package name with one that does not
include the major version number.

Changelog: other