Makefile: Fix protoc linting not working
With 2a333822f (Use protoc_gen_gitaly plugin when regenerating protobuf
definitions, 2022-04-29), we have merged usage of Gitaly's protoc plugin
into the use of `make proto`. This was done because the plugin doesn't
only lint sources, but it also generates some data structures on its
own.

This process has broken use of the plugin completely though. While we
make the plugin's path known to `protoc` via the `--plugin` option, it
is not getting used at all because we didn't also move `--gitaly_out`.
This option is required though for the plugin to actually be used.

Fix this bug by adding the option back in.