git: Move fallback time into `catfile` package
When we fail to parse the commit time in the `catfile` package, then we
instead return a fallback time. While the `catfile` package is the only
user of this value, it is currently contained in the `git` package as a
public variable.

Move it over into the `catfile` package to keep it close to its user.