git: Rename file hosting revision verification logic
The logic to verify that revisions are valid revisions that can safely
be passed to various Git components is hosted in a file `proto.go`. This
is probably a historic artifact, but it doesn't make a lot of sense
nowadays.

Rename the file to `revision.go` to make it easier to find. While at it,
this commit also moves related tests into `revision_test.go`.