Return correct error when repository has no up to date replicas
GetConsistentStorages retrieves the up to date replicas of a repository.
If the repository has no up to date replicas, it returns a repository
not found error. This is incorrect as the repository may still exist
from Praefect's point of view, it's just that the replicas with the
latest changes have been lost. This commit changes GetConsistentStorages
to return a more descriptive error for this scenario instead of a
repository not found error.

Changelog: fixed