Update gitaly package name by removing major version number
Since Gitaly doesn't have an extensive public facing API and is not
meant to be used as such, including the major version number in the
package name doesn't help. Rather since we release a major version
regularly, this ends up being a lot of maintenance work to bump the
version number in the package name each time.

Remove the major version name from the package name in go files.

Changelog: other