gitpipe: Add tests demonstrating loss of context-cancellation errors
We're not bubbling up any errors in case contexts get cancelled while we
were consuming pipeline result. This error makes it easy to think that
the pipeline finished successfully even though the pipeline has been
cancelled and only returned partial results.

Add a set of tests to demonstrate this behaviour.