go.mod: Remove exclude directive
Remove the exlude directive that prevented us from pulling in a a buggy
version of grpc/grpc-go. We don't use this package anymore, so there's
no reason to keep this directive. Also, newer versions of grpc/grpc-go
has fixed the bug that's mentioned as the reason for the exclude.

Changelog: other